api/requirements.txt
backcall==0.2.0
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flask-mail==0.9.1
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
mongoengine==0.20.0
numpy==1.19.1
pandas==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
psycopg2==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
traitlets==4.3.3
urllib3==1.24.3
uWSGI==2.0.19.1
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==1.2